pyproject.toml
[tool.poetry]
name = "drf-homework1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.3.0"
djangorestframework = "^3.15.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
django = "^5.0.6"
coverage = "^7.5.3"
drf-yasg = "^1.21.7"
setuptools = "^70.0.0"
stripe = "^9.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
